Contributing community 'works on these' list
I've successfully been using v6.12 on an AzDelivery ESP32 Dev Kit C ESP32-WROOM-32 v2 (version 2) for the last month or so with zero issues. Thought I'd contribute to say so.

I've just swapped over to a v4 of the same board, and again appear to be experiencing no issues.

Both were deployed with the `az-delivery-devkit-v4` profile in platformio and the `LED_BUILTIN` set to 1

I'll upgrade to 7.0+ in the coming weeks and report back then :)

No drama if each individual dev board is not a useful contribution; but with the variance between dev boards sometimes when you're a noob (like I was!) you need to see specifics to know whether something will work.